Fix VTs hash check and add --dump-vt-verification
The collation of the VTs and preferences is now set explicitly in the
SQL used to get the string the VTs verification hash is calculated from
to ensure the VTs and preferences are sorted consistently.
The unhashed string is now also generated in a function and the command
line option --dump-vt-verification has been added.

(cherry picked from commit fc2cf6f9eec7440b56b07e4461aebdf1978438bf)